--- frameworks/Swift/vapor/Sources/TfbCommon/Message.swift
@@ -1,9 +1,0 @@
-import Vapor
-
-public struct Message: Content {
-    public let message: String
-
-    public init(_ message: String) {
-        self.message = message
-    }
-}
