--- frameworks/Swift/vapor/Sources/vapor-tfb/boot.swift
@@ -1,7 +1,0 @@
-import Routing
-import Vapor
-
-/// Called after your application has initialized.
-///
-/// [Learn More →](https://docs.vapor.codes/3.0/getting-started/structure/#bootswift)
-public func boot(_ app: Application) throws {}
