--- frameworks/Swift/vapor/Sources/vapor-tfb/config.swift
@@ -1,25 +1,0 @@
-import Vapor
-import TfbCommon
-
-/// Called before your application initializes.
-///
-/// [Learn More →](https://docs.vapor.codes/3.0/getting-started/structure/#configureswift)
-public func configure(
-    _ config: inout Config,
-    _ env: inout Environment,
-    _ services: inout Services
-) throws {
-    services.register(ServerMiddleware.self)
-
-    // Register routes to the router
-    let router = EngineRouter.default()
-    try routes(router)
-    services.register(router, as: Router.self)
-
-    // Register middleware
-    var middlewares = MiddlewareConfig()
-    middlewares.use(ServerMiddleware.self)
-    services.register(middlewares)
-
-    // Configure the rest of your application here
-}
