+++ frameworks/Swift/vapor/Sources/vapor-tfb/main.swift
@@ -1,25 +1,28 @@
-import Service
 import Vapor
-import Foundation
 
-// The contents of main are wrapped in a do/catch block because any errors that get raised to the top level will crash Xcode
-do {
-    var config = Config.default()
-    var env = try Environment.detect()
-    var services = Services.default()
-
-    try configure(&config, &env, &services)
-
-    let app = try Application(
-        config: config,
-        environment: env,
-        services: services
-    )
-
-    try boot(app)
-
-    try app.run()
-} catch {
-    print(error)
-    exit(1)
+// Models
+struct Message: Content {
+	let message: String?
 }
+
+// Services
+var services = Services.default()
+// services.register(ServerMiddleware.self)
+
+// Routes
+let router = EngineRouter.default()
+router.get("json") { req in
+    return Message(message: nil)
+}
+router.get("plaintext") { req in
+    return "Hello, world!" as StaticString
+}
+services.register(router, as: Router.self)
+
+// Middlewares
+var middlewares = MiddlewareConfig()
+//middlewares.use(ServerMiddleware.self)
+services.register(middlewares)
+
+let app = try Application(config: .default(), environment: .detect(), services: services)
+try app.run()
